feat: 報告書兼請求書の明細を刷新(小数点入力・列順=案件/種別/数量/単価/合計・単位枚/一式)+DL明確化
- posting_logic: _num/unit_for/fmt_num を追加、集計を小数対応(int→_num)、金額表示は.0を出さないfmt_num
- pages/02: 明細エディタ列順を案件・種別・数量・単価に、数量/単価をNumberColumn(小数可 %g)、
  明細(確認)プレビューを案件・種別・数量(+単位 枚/一式)・単価・合計の順で表示
- invoice_excel: 数量/金額を小数保持、請求額・部数はfmt_num
- posting_store: 委託明細の数量/単価をfloat保存
- #18: DLは現行st.download_buttonで正常(日本語ファイル名でもattachmentでDL成功を実機確認)。ボタン文言を明確化しkey付与

+++ common/posting_store.py
@@ -363,8 +363,8 @@ def add_contract_invoice(distributor_id, issue_date, period_from, period_to, lin
             (_int_or_none(distributor_id), issue_date, period_from, period_to, _now(now)))
         invoice_id = int(cur.lastrowid)
         for ln in lines:
-            qty = int(ln.get("report_qty") or 0)
-            price = int(ln.get("unit_price") or 0)
+            qty = float(ln.get("report_qty") or 0)
+            price = float(ln.get("unit_price") or 0)
             conn.execute(
                 "INSERT INTO contract_invoice_lines"
                 " (invoice_id, project_id, report_qty, unit_price, amount, remark)"
